+++ internal/tests/ApplicationDeploymentResource/application_deployment_resource_test.go
@@ -46,3 +46,46 @@ func TestApplicationDeploymentResource(t *testing.T) {
 		},
 	})
 }
+
+func TestApplicationDeploymentKSMLResource(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: GetProviderConfig(t).ProtoV6ProviderFactories,
+		ExternalProviders:        GetProviderConfig(t).ExternalProviders,
+
+		Steps: []resource.TestStep{
+			{
+				Config: GetProvider() + GetFile("axual_application_deployment_ksml_initial.tf"),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrPair("axual_application_deployment.ksml_axual_application_deployment", "environment", "axual_environment.tf-test-ksml-env", "id"),
+					resource.TestCheckResourceAttrPair("axual_application_deployment.ksml_axual_application_deployment", "application", "axual_application.tf-test-ksml-app", "id"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "type", "KSML"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "deployment_size", "S"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "restart_policy", "on_exit"),
+					resource.TestCheckResourceAttrSet("axual_application_deployment.ksml_axual_application_deployment", "definition"),
+				),
+			},
+			{
+				Config: GetProvider() + GetFile("axual_application_deployment_ksml_updated.tf"),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrPair("axual_application_deployment.ksml_axual_application_deployment", "environment", "axual_environment.tf-test-ksml-env", "id"),
+					resource.TestCheckResourceAttrPair("axual_application_deployment.ksml_axual_application_deployment", "application", "axual_application.tf-test-ksml-app", "id"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "type", "KSML"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "deployment_size", "M"),
+					resource.TestCheckResourceAttr("axual_application_deployment.ksml_axual_application_deployment", "restart_policy", "on_exit"),
+					resource.TestCheckResourceAttrSet("axual_application_deployment.ksml_axual_application_deployment", "definition"),
+				),
+			},
+			{
+				ResourceName:         "axual_application_deployment.ksml_axual_application_deployment",
+				ImportState:          true,
+				ImportStateVerify:    true,
+				Config:               GetProvider() + GetFile("axual_application_deployment_ksml_updated.tf"),
+			},
+			{
+				// To ensure cleanup if one of the test cases had an error
+				Destroy: true,
+				Config:  GetProvider() + GetFile("axual_application_deployment_ksml_updated.tf"),
+			},
+		},
+	})
+}